fix: improve payload validation and documentation clarity
Agent-Logs-Url: https://github.com/deogratius2phantom/Indoor-Positioning-System/sessions/faa4297f-2002-403a-a654-62b5af10f7c2

+++ firmware/esp32_sniffer/esp32_sniffer.ino
@@ -146,7 +146,7 @@ void loop() {
   }
 
   const uint32_t now = millis();
-  if (now - lastHop >= kHopIntervalMs) {
+  if (static_cast<uint32_t>(now - lastHop) >= kHopIntervalMs) {
     currentChannel = (currentChannel >= kMaxChannel) ? kMinChannel : (currentChannel + 1);
     esp_wifi_set_channel(currentChannel, WIFI_SECOND_CHAN_NONE);
     lastHop = now;
